requirements.txt
requests~=2.32
beautifulsoup4~=4.12
pandas~=2.2
python-dotenv
sqlalchemy
psycopg2-binary